server/requirement.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
google-api-python-client>=2.75.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
youtube-transcript-api>=0.6.0
transformers>=4.30.0
torch>=2.0.0
google-generativeai>=0.3.0
pyparsing>=3.0.0
protobuf>=4.23.0
numpy>=1.24.0
tqdm>=4.65.0
sentencepiece>=0.1.99
accelerate>=0.22.0